Adding cache teset

+++ src/main/java/dev/charlesardsilva/lecturecachecontent/service/ReviewService.java
@@ -1,5 +1,5 @@
 package dev.charlesardsilva.lecturecachecontent.service;
 
 public interface ReviewService {
-    int rating(String id);
+    double rating(String id);
 }
